drivers: sensor: icm42605: added i2c support
The icm42605 has hardware support for both spi and i2c interfaces, however the driver only supports the spi bus. Adding driver support for i2c interface.

+++ drivers/sensor/tdk/icm42605/icm42605_setup.h
@@ -14,6 +14,6 @@ int icm42605_turn_on_fifo(const struct device *dev);
 int icm42605_turn_off_fifo(const struct device *dev);
 int icm42605_turn_off_sensor(const struct device *dev);
 int icm42605_turn_on_sensor(const struct device *dev);
-int icm42605_set_odr(const struct device *dev, int a_rate, int g_rate);
+int icm42605_set_odr(const struct device *dev, uint16_t a_rate, uint16_t g_rate);
 
 #endif /* __SENSOR_ICM42605_ICM42605_SETUP__ */
